Optimize backtraces.
